extract QP construction code out of MPCController and create MPCQPFactor
